PLANET-5409 promote in compare job
* The promotion used to be in a seprate job, however doing it in compare
directly is a lot faster as 1 less container needs to spin up.
* Because promotion happens inside the job now, there is no more reason
to make it fail if there's a hold prefix. Instead the hold prefix will
also trigger a rollback workflow, as in that case we expect changes
might be needed, so we prepare the rollback for should it be needed.