Use specific selectors for default scenarios
For the only two default scenarios we run on all NROs, we can switch
on using specific selectors, instead of testing the whole page, to avoid false negatives.

For NROs that have heavily customized child themes, we can still add the full page
scenarios in their scenarios. Since we use these tests in deployment as a sanity check,
for NROs that use the vanilla theme checking specific selectors is sufficient.
And we already have acceptance tests to check specific blocks.